Projects2026/Level2/unsupervised-learning/requirements.txt
numpy
matplotlib
scikit-learn
umap-learn
